+++ src/daceInterface.ts
@@ -1,9 +1,8 @@
 // Copyright 2020-2021 ETH Zurich and the DaCe-VSCode authors.
 // All rights reserved.
 
-import * as path from 'path';
+import * as os from 'os';
 import * as vscode from 'vscode';
-import * as cp from 'child_process';
 import { request } from 'http';
 
 import { DaCeVSCode } from './extension';
@@ -64,6 +63,9 @@ implements MessageReceiverInterface {
                 if (message.name)
                     this.getEnum(message.name, origin);
                 break;
+            case 'query_sdfg_metadata':
+                this.querySdfgMetadata();
+                break;
             default:
                 break;
         }
@@ -74,6 +76,7 @@ implements MessageReceiverInterface {
     }
 
     private runTerminal?: vscode.Terminal = undefined;
+    private daemonTerminal?: vscode.Terminal = undefined;
 
     private daemonRunning = false;
     private daemonBooting = false;
@@ -82,29 +85,52 @@ implements MessageReceiverInterface {
         'dace.interface'
     ).port;
 
-    public async getPythonPath(
-        document: vscode.TextDocument | null
+    public async getPythonExecCommand(
+        uri: vscode.Uri | undefined
     ): Promise<string> {
         try {
             let pyExt = vscode.extensions.getExtension('ms-python.python');
-            if (!pyExt)
+            if (!pyExt) {
+                // TODO: do we want to tell the user that using the python
+                // plugin might be advisable here?
                 return 'python';
+            }
 
             if (pyExt.packageJSON?.featureFlags?.usingNewInterpreterStorage) {
                 if (!pyExt.isActive)
                     await pyExt.activate();
-                const pythonPath = pyExt.exports.settings.getExecutionDetails ?
+                const pyCmd = pyExt.exports.settings.getExecutionDetails ?
                     pyExt.exports.settings.getExecutionDetails(
-                        document?.uri
+                        uri
                     ).execCommand :
-                    pyExt.exports.settings.getExecutionCommand(document?.uri);
-                return pythonPath ? pythonPath.join(' ') : 'python';
+                    pyExt.exports.settings.getExecutionCommand(uri);
+                if (pyCmd) {
+                    // Ensure spaces in the python command don't trip up the
+                    // terminal.
+                    switch (os.platform()) {
+                        case 'win32':
+                            for (let i = 0; i < pyCmd.length; i++) {
+                                if (/\s/g.test(pyCmd[i]))
+                                    pyCmd[i] = '& "' + pyCmd[i] + '"';
+                            }
+                            break;
+                        default:
+                            for (let i = 0; i < pyCmd.length; i++) {
+                                if (/\s/g.test(pyCmd[i]))
+                                    pyCmd[i] = '"' + pyCmd[i] + '"';
+                            }
+                            break;
+                    }
+                    return pyCmd.join(' ');
+                } else {
+                    return 'python';
+                }
             } else {
                 let path = undefined;
-                if (document)
+                if (uri)
                     path = vscode.workspace.getConfiguration(
                         'python',
-                        document.uri
+                        uri
                     ).get<string>('pythonPath');
                 else
                     path = vscode.workspace.getConfiguration(
@@ -134,47 +160,33 @@ implements MessageReceiverInterface {
         }
     }
 
-    private genericBackendErrorPopup() {
-        vscode.window.showErrorMessage(
-            'Encountered an error in the DaCe daemon! ',
-            'Show Error Output',
-            'Retry in Terminal Mode'
-        ).then((opt) => {
-            switch (opt) {
-                case 'Show Error Output':
-                    DaCeVSCode.getInstance().getOutputChannel().show();
-                    break;
-                case 'Retry in Terminal Mode':
-                    vscode.commands.executeCommand(
-                        'dace.openOptimizerInTerminal'
-                    );
-                    break;
-            }
-        });
-    }
-
-    private getRunDaceScriptPath(): string | undefined{
-        const extensionPath =
-            DaCeVSCode.getInstance().getExtensionContext()?.extensionPath;
-        if (!extensionPath) {
-            DaCeVSCode.getInstance().getOutputChannel().append(
+    private getRunDaceScriptUri(): vscode.Uri | undefined{
+        const extensionUri =
+            DaCeVSCode.getInstance().getExtensionContext()?.extensionUri;
+        if (!extensionUri) {
+            vscode.window.showErrorMessage(
                 'Failed to load the file path to the extension'
             );
-            this.genericBackendErrorPopup();
             return undefined;
         }
-        return path.join(
-            extensionPath, 'backend', 'run_dace.py'
-        );
+        return vscode.Uri.joinPath(extensionUri, 'backend', 'run_dace.py');
     }
 
-    public startDaemonInTerminal(callback?: CallableFunction) {
-        const term = vscode.window.createTerminal('SDFG Optimizer');
-        term.show();
-        const scriptPath = this.getRunDaceScriptPath();
-        if (scriptPath) {
-            term.sendText(
-                'python ' + scriptPath + ' -p ' + this.port.toString()
+    public async startDaemonInTerminal(callback?: CallableFunction) {
+        if (this.daemonTerminal === undefined)
+            this.daemonTerminal = vscode.window.createTerminal(
+                'SDFG Optimizer'
+            );
+
+        const scriptUri = this.getRunDaceScriptUri();
+        if (scriptUri) {
+            vscode.window.setStatusBarMessage(
+                'Trying to start and connect to a DaCe daemon', 5000
+            );
+            const pyCmd: string = await this.getPythonExecCommand(scriptUri);
+
+            this.daemonTerminal?.sendText(
+                pyCmd + ' ' + scriptUri.fsPath + ' -p ' + this.port.toString()
             );
             this.pollDaemon(callback, true);
         } else {
@@ -324,22 +336,12 @@ implements MessageReceiverInterface {
                     'Unable to start and connect to DaCe. Do you have it ' +
                     'installed?',
                     'Retry',
-                    'Retry in Terminal Mode',
                     'Install DaCe'
                 ).then(opt => {
                     switch (opt) {
                         case 'Retry':
                             clearInterval(connectionIntervalId);
-                            this.startPythonDaemon();
-                            break;
-                        case 'Retry in Terminal Mode':
-                            vscode.commands.executeCommand(
-                                'dace.openOptimizerInTerminal'
-                            );
-                            // Do not clear the connection interval immediately
-                            setTimeout(() => {
-                                clearInterval(connectionIntervalId);
-                            }, 10000);
+                            this.startDaemonInTerminal();
                             break;
                         case 'Install DaCe':
                             clearInterval(connectionIntervalId);
@@ -351,49 +353,6 @@ implements MessageReceiverInterface {
                 clearInterval(connectionIntervalId);
             }
         }, 10000);
-    }
-
-    private async startPythonDaemon(callback?: CallableFunction) {
-        if (this.daemonRunning) {
-            if (callback)
-                callback();
-            return;
-        }
-
-        vscode.window.setStatusBarMessage(
-            'Trying to start and connect to a DaCe daemon', 5000
-        );
-
-        const pythonPath = await this.getPythonPath(null);
-        const scriptPath = this.getRunDaceScriptPath();
-        if (!scriptPath)
-            return;
-
-        const workspaceFolders = vscode.workspace.workspaceFolders;
-        let workspaceRoot = undefined;
-        if (workspaceFolders)
-            workspaceRoot = workspaceFolders[0].uri.fsPath;
-
-        const daemon = cp.spawn(
-            pythonPath,
-            [scriptPath, '-p', this.port.toString()], {
-                cwd: workspaceRoot,
-            }
-        );
-
-        daemon.on('exit', (_code, _signal) => {
-            this.daemonRunning = false;
-            this.daemonBooting = false;
-        });
-
-        daemon.stderr.on('data', data => {
-            DaCeVSCode.getInstance().getOutputChannel().append(
-                data.toString()
-            );
-            this.genericBackendErrorPopup();
-        });
-
-        this.pollDaemon(callback, false);
     }
 
     private sendRequest(url: string,
@@ -515,7 +474,7 @@ implements MessageReceiverInterface {
         ).then(opt => {
             switch (opt) {
                 case 'Yes':
-                    DaCeInterface.getInstance().startPythonDaemon();
+                    DaCeInterface.getInstance().startDaemonInTerminal();
                     break;
                 case 'No':
                     break;
@@ -535,15 +494,10 @@ implements MessageReceiverInterface {
 
         const callback = () => {
             TransformationHistoryProvider.getInstance()?.refresh();
-            TransformationListProvider.getInstance()?.refresh();
+            TransformationListProvider.getInstance()?.refresh(true);
+            this.querySdfgMetadata();
         };
-        if (vscode.workspace.getConfiguration(
-                'dace.interface'
-            ).terminalMode === true
-        )
-            this.startDaemonInTerminal(callback);
-        else
-            this.startPythonDaemon(callback);
+        this.startDaemonInTerminal(callback);
     }
 
     public previewSdfg(sdfg: any, history_mode: boolean = false) {
@@ -758,6 +712,21 @@ implements MessageReceiverInterface {
             },
             callback
         );
+    }
+
+    public async querySdfgMetadata() {
+        async function callback(data: any) {
+            SdfgViewerProvider.getInstance()?.handleMessage({
+                type: 'set_sdfg_metadata',
+                meta_dict: data.meta_dict,
+            });
+        };
+
+        if (this.daemonRunning)
+            this.sendGetRequest(
+                '/get_metadata',
+                callback
+            );
     }
 
     public async loadTransformations(sdfg: any, selectedElements: any) {